ml-service/requirements.txt
flask==3.0.0
flask-cors==4.0.0
Pillow==10.2.0
numpy==1.26.3
scipy==1.12.0
